refactor to general export code module
