+++ src/components/export-code/ExportableCode.ts
@@ -1,3 +1,5 @@
+// ~~~~~~~~~ EXPORT CHART CODE ~~~~~~~~~
+
 const VIS_EXPORTS = `# Generated by AutoProfiler
 
 import altair as alt`
